requirements.txt
ipykernel 
telethon
python-dotenv